django_orm/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
Baker==1.3
flake8==4.0.1
model-bakery==1.6.0
psycopg2-binary==2.9.3